// api-legacy/src/core/Entity/Council_Entity_Document.gen.ts
/* TypeScript file generated from Council_Entity_Document.res by genType. */
/* eslint-disable import/first */


// @ts-ignore: Implicit any on import
const Curry = require('rescript/lib/js/curry.js');

// @ts-ignore: Implicit any on import
const Council_Entity_DocumentBS = require('./Council_Entity_Document.bs');

import type {t as Date_t} from '../../../src/core/Date.gen';

// tslint:disable-next-line:interface-over-type-literal
export type id = string;

// tslint:disable-next-line:interface-over-type-literal
export type sectionId = string;

// tslint:disable-next-line:interface-over-type-literal
export type memberId = string;

// tslint:disable-next-line:interface-over-type-literal
export type data = {
  readonly title: string; 
  readonly sections: sectionId[]; 
  readonly tags: string[]; 
  readonly owner: memberId; 
  readonly responsibility: memberId
};

// tslint:disable-next-line:interface-over-type-literal
export type state = 
    { tag: "Free"; value: data }
  | { tag: "Locked"; value: { readonly by: memberId; readonly data: data } };

// tslint:disable-next-line:interface-over-type-literal
export type t = {
  readonly _RE: 
    "Document"; 
  readonly id: id; 
  readonly seq: number; 
  readonly state?: state
};

// tslint:disable-next-line:interface-over-type-literal
export type event = 
    { tag: "Created"; value: { readonly date: Date_t; readonly data: data } }
  | { tag: "SectionAdded"; value: { readonly date: Date_t; readonly section: sectionId } }
  | { tag: "SectionDeleted"; value: { readonly date: Date_t; readonly section: sectionId } }
  | { tag: "ResponsibilityAssigned"; value: { readonly date: Date_t; readonly responsibility: memberId } }
  | { tag: "TagsModified"; value: { readonly date: Date_t; readonly tags: string[] } }
  | { tag: "Locked"; value: { readonly date: Date_t; readonly by: memberId } };

// tslint:disable-next-line:interface-over-type-literal
export type error = 
    { tag: "Uninitialized"; value: id }
  | { tag: "Locked"; value: id };

export const make: (id:id, _2:{ readonly state?: state; readonly seq?: number }, _3:void) => t = function (Arg1: any, Arg2: any, Arg3: any) {
  const result = Curry._4(Council_Entity_DocumentBS.make, Arg1, (Arg2.state == null ? undefined : Arg2.state.tag==="Free"
    ? {TAG: 0, _0:Arg2.state.value} as any
    : Object.assign({TAG: 1}, Arg2.state.value)), Arg2.seq, Arg3);
  return {_RE:result._RE, id:result.id, seq:result.seq, state:(result.state == null ? result.state : result.state.TAG===0
    ? {tag:"Free", value:result.state._0}
    : {tag:"Locked", value:result.state})}
};